python_service/requirements.txt
osascript==2020.12.3
pyserial==3.5

